requirements.txt
fastapi
uvicorn[standard]
httpx
qdrant-client
pydantic-settings
python-dotenv
pypdf
python-docx
python-multipart